+++ dashboard/src/hooks/useFiles.ts
@@ -1,4 +1,5 @@
 import generateAppServiceUrl from '@/utils/common/generateAppServiceUrl';
+import { getHasuraAdminSecret } from '@/utils/env';
 import type {
   Files_Order_By as FilesOrderBy,
   GetFilesQuery,
@@ -92,7 +93,7 @@ export default function useFiles({
                     headers: {
                       'x-hasura-admin-secret':
                         process.env.NEXT_PUBLIC_ENV === 'dev'
-                          ? 'nhost-admin-secret'
+                          ? getHasuraAdminSecret()
                           : currentApplication?.config?.hasura.adminSecret,
                     },
                     mode: 'cors',