+++ dashboard/src/utils/nhost.ts
@@ -1,9 +1,25 @@
+import {
+  getAuthServiceUrl,
+  getFunctionsServiceUrl,
+  getGraphqlServiceUrl,
+  getStorageServiceUrl,
+  isPlatform,
+} from '@/utils/env';
 import { NhostClient } from '@nhost/nextjs';
-import { LOCAL_SUBDOMAIN } from './env';
 
-export const nhost =
-  process.env.NEXT_PUBLIC_NHOST_PLATFORM === 'true'
-    ? new NhostClient({ backendUrl: process.env.NEXT_PUBLIC_NHOST_BACKEND_URL })
-    : new NhostClient({ subdomain: LOCAL_SUBDOMAIN });
+// eslint-disable-next-line no-nested-ternary
+export const nhost = isPlatform()
+  ? new NhostClient({ backendUrl: process.env.NEXT_PUBLIC_NHOST_BACKEND_URL })
+  : getAuthServiceUrl() &&
+    getGraphqlServiceUrl() &&
+    getStorageServiceUrl() &&
+    getFunctionsServiceUrl()
+  ? new NhostClient({
+      authUrl: getAuthServiceUrl(),
+      graphqlUrl: getGraphqlServiceUrl(),
+      storageUrl: getStorageServiceUrl(),
+      functionsUrl: getFunctionsServiceUrl(),
+    })
+  : new NhostClient({ subdomain: 'local' });
 
 export default nhost;